+++ app/src/main/java/ru/drsk/progserega/inspectionsheet/entities/organization/ElectricNetworkArea.java
@@ -4,11 +4,11 @@ package ru.drsk.progserega.inspectionsheet.entities.organization;
  * Район Электрических Сетей
  */
 public class ElectricNetworkArea {
-    private int id;
+    private long id;
     private String name;
     private NetworkEnterprise networkEnterprise;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -20,7 +20,7 @@ public class ElectricNetworkArea {
         return networkEnterprise;
     }
 
-    public ElectricNetworkArea(int id, String name, NetworkEnterprise networkEnterprise) {
+    public ElectricNetworkArea(long id, String name, NetworkEnterprise networkEnterprise) {
         this.id = id;
         this.name = name;
         this.networkEnterprise = networkEnterprise;
